+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+# v0.4.0
+
+- Add 10 new utilities
+
+# v0.3.0
+
+- Fix support of string as arguments for utilities
+- Add Size utility
+- Add Sticky Footer utility
+- Add Word Wrap utility
+
+# v0.2.1
+
+- Add support of function as arguments for utilities
+
 # v0.2.0
 
 - Public release